♻️ refactor:예약 삭제 api 연결
Issue Resolved: #159
